requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
sqlalchemy>=2.0
pydantic>=2.5
fpdf2>=2.7
python-dotenv>=1.0
